# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sito DNS Administrative API",
    "description": "Administrative REST API and OpenAPI 3.0 specification for the sito DNS server",
    "license": {
      "name": "GPL-3.0-only"
    },
    "version": "1.6.0"
  },
  "paths": {
...
    "/api/v1/filtering/check": {
      "post": {
        "tags": [
          "Filtering"
        ],
        "operationId": "check_filtering",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filtering check result simulated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/filtering/lists": {
      "get": {
        "tags": [
          "Filtering"
        ],
        "operationId": "get_filter_lists",
        "responses": {
          "200": {
            "description": "Subscription lists retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilterListDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Filtering"
        ],
        "operationId": "add_filter_list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFilterListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filter list added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterListDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/filtering/lists/{id}": {
      "put": {
        "tags": [
          "Filtering"
        ],
        "operationId": "update_filter_list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Filter list index ID",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFilterListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filter list updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterListDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Filtering"
        ],
        "operationId": "delete_filter_list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Filter list index ID",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filter list deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/filtering/refresh": {
      "post": {
        "tags": [
          "Filtering"
        ],
        "operationId": "refresh_filtering",
        "responses": {
          "200": {
            "description": "Filter lists refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/filtering/rules": {
      "get": {
        "tags": [
          "Filtering"
        ],
        "operationId": "get_filtering_rules",
        "responses": {
          "200": {
            "description": "Custom rules retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRulesDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Filtering"
        ],
        "operationId": "set_filtering_rules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRulesDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Custom rules updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRulesDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddFilterListRequest": {
        "type": "object",
        "description": "Request to create a new filter list.",
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "refresh_hours": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "CustomRulesDto": {
        "type": "object",
        "description": "Custom rules collection.",
        "required": [
          "rules"
        ],
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FilterCheckRequest": {
        "type": "object",
        "description": "Request to test/simulate a filtering verdict without contacting upstreams.",
        "required": [
          "domain"
        ],
        "properties": {
          "client": {
            "type": [
              "string",
              "null"
            ]
          },
          "domain": {
            "type": "string"
          },
          "qtype": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "FilterCheckResponse": {
        "type": "object",
        "description": "Simulated filtering verdict response.",
        "required": [
          "domain",
          "verdict"
        ],
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "domain": {
            "type": "string"
          },
          "list_source": {
            "type": [
              "string",
              "null"
            ]
          },
          "rule": {
            "type": [
              "string",
              "null"
            ]
          },
          "verdict": {
            "type": "string"
          }
        }
      },
      "FilterListDto": {
        "type": "object",
        "description": "Filter list subscription details. `id` is the stable list name (numeric\nindices from older clients are still accepted on update/delete).",
        "required": [
          "id",
          "name",
          "url",
          "enabled",
          "refresh_hours",
          "rule_count"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "last_updated": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "refresh_hours": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "rule_count": {
            "type": "integer",
            "minimum": 0
          },
          "url": {
            "type": "string"
          }
        }
      },
      "GenericMessageResponse": {
        "type": "object",
        "description": "Generic success message response.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateFilterListRequest": {
        "type": "object",
        "description": "Request to update an existing filter list.",
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_hours": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}